Feat: #166 User와 Auth 모듈 분리
