+++ src/modules/user/user.repository.ts
@@ -14,26 +14,6 @@ import { GetMakerListQueryDTO } from 'src/common/types/user/query.dto';
 export default class UserRepository {
   constructor(private readonly db: DBClient) {}
 
-  async findByEmail(email: string): Promise<IUser> {
-    const data = await this.db.user.findUnique({
-      where: {
-        email
-      }
-    });
-
-    return new UserMapper(data).toDomain();
-  }
-
-  async findByNickName(nickName: string): Promise<IUser> {
-    const data = await this.db.user.findUnique({
-      where: {
-        nickName
-      }
-    });
-
-    return new UserMapper(data).toDomain();
-  }
-
   async findById(id: string): Promise<IUser> {
     const data = await this.db.user.findUnique({ where: { id } });
 
@@ -108,20 +88,6 @@ export default class UserRepository {
     });
   }
 
-  async create(user: Partial<UserProperties>): Promise<IUser> {
-    const data = await this.db.user.create({
-      data: {
-        role: user.role,
-        email: user.email,
-        nickName: user.nickName,
-        password: user.password,
-        phoneNumber: user.phoneNumber
-      }
-    });
-
-    return new UserMapper(data).toDomain();
-  }
-
   async update(id: string, data: Partial<UserProperties>): Promise<IUser> {
     const user = await this.db.user.update({
       where: {
@@ -131,35 +97,6 @@ export default class UserRepository {
     });
 
     return new UserMapper(user).toDomain();
-  }
-
-  async createDreamer(user: Partial<DreamerProfileProperties>): Promise<IDreamerProfile> {
-    const profile = await this.db.dreamerProfile.create({
-      data: {
-        user: { connect: { id: user.userId } },
-        tripTypes: user.tripTypes,
-        serviceArea: user.serviceArea,
-        image: user.image
-      }
-    });
-
-    return new DreamerProfileMapper(profile).toDomain();
-  }
-
-  async createMaker(user: Partial<MakerProfileProperties>): Promise<IMakerProfile> {
-    const profile = await this.db.makerProfile.create({
-      data: {
-        user: { connect: { id: user.userId } },
-        serviceArea: user.serviceArea,
-        serviceTypes: user.serviceTypes,
-        gallery: user.gallery,
-        description: user.description,
-        detailDescription: user.detailDescription,
-        image: user.image
-      }
-    });
-
-    return new MakerProfileMapper(profile).toDomain();
   }
 
   async findDreamerProfile(userId: string): Promise<IDreamerProfile> {
