+++ src/modules/user/user.service.ts
@@ -2,9 +2,6 @@ import { Injectable } from '@nestjs/common';
 import UserRepository from './user.repository';
 import BadRequestError from 'src/common/errors/badRequestError';
 import ErrorMessage from 'src/common/constants/errorMessage.enum';
-import User from '../../common/domains/user/user.domain';
-import { JwtService } from '@nestjs/jwt';
-import { DreamerProfile, MakerProfile } from '../../common/domains/user/profile.domain';
 import { FilteredUserProperties, PasswordProperties, UserProperties } from '../../common/types/user/user.types';
 import {
   DreamerProfileProperties,
@@ -20,58 +17,9 @@ import { followResponseDTO, ProfileCardResponseDTO } from 'src/common/types/user
 export default class UserService {
   constructor(
     private readonly repository: UserRepository,
-    private readonly jwt: JwtService,
     private readonly userStats: UserStatsService,
     private readonly follow: FollowService
   ) {}
-
-  async createUser(data): Promise<null> {
-    const { user, profile } = data;
-
-    // FIXME: transaction utility 만들어서 적용하자
-    // 유저 등록
-    const existingEmail = await this.repository.findByEmail(user.email);
-    if (existingEmail) {
-      throw new BadRequestError(ErrorMessage.USER_EXIST);
-    }
-
-    const existingNickName = await this.repository.findByNickName(user.nickName);
-    if (existingNickName) {
-      throw new BadRequestError(ErrorMessage.USER_NICKNAME_EXIST);
-    }
-
-    const userData = await User.create(user);
-    const savedUser = await this.repository.create(userData.get());
-    const newUser = savedUser.get();
-
-    // 역할에 따라 프로필 등록
-    if (newUser.role === 'DREAMER') {
-      const profileData = DreamerProfile.create({ ...profile, userId: newUser.id });
-      await this.repository.createDreamer(profileData);
-    } else {
-      const profileData = MakerProfile.create({ ...profile, userId: newUser.id });
-      await this.repository.createMaker(profileData.get());
-    }
-
-    // 유저 생성시 기본값으로 UserStats 생성
-    await this.userStats.create(savedUser.getId(), {});
-
-    return;
-  }
-
-  async login(email: string, password: string): Promise<FilteredUserProperties> {
-    const user = await this.repository.findByEmail(email);
-    if (!user) {
-      throw new BadRequestError(ErrorMessage.USER_UNAUTHORIZED_ID);
-    }
-
-    const isValidPassword = await user.validatePassword(password);
-    if (!isValidPassword) {
-      throw new BadRequestError(ErrorMessage.USER_BAD_REQUEST_PW);
-    }
-
-    return user.toClient();
-  }
 
   async getUser(userId: string): Promise<Omit<UserProperties, 'password'>> {
     const user = await this.repository.findById(userId);
@@ -97,18 +45,6 @@ export default class UserService {
     const totalCount = await this.repository.count(options);
 
     return { totalCount, list };
-  }
-
-  createTokens(payload: { userId: string; role: string }) {
-    const accessToken = this.jwt.sign(payload, { expiresIn: process.env.ACCESS_TOKEN_EXPIRY });
-    const refreshToken = this.jwt.sign(payload, { expiresIn: process.env.REFRESH_TOKEN_EXPIRY });
-
-    return { accessToken, refreshToken };
-  }
-
-  createNewToken(oldToken: string) {
-    const { userId, role } = this.jwt.verify(oldToken);
-    return this.createTokens({ userId, role });
   }
 
   async updateUser(
@@ -146,18 +82,6 @@ export default class UserService {
 
     const newProfile = await this.repository.updateMakerProfile(userId, profile.update(data));
     return newProfile.get();
-  }
-
-  async checkEmail(email: string): Promise<boolean> {
-    const user = await this.repository.findByEmail(email);
-
-    return !user;
-  }
-
-  async checkNickName(nickName: string): Promise<boolean> {
-    const user = await this.repository.findByNickName(nickName);
-
-    return !user;
   }
 
   async getProfileCardData(makerId: string, dreamerId: string, withDetails?: boolean): Promise<ProfileCardResponseDTO> {
